agno-demo/requirements.txt
agno>=2.0.0
groq
python-dotenv